Debut intégration homepage
BUGFIX : supression de savoir qui plante si l'image n'existe pas
BUGFIX : affichage exercices dans l'epreuve

+++ src/Main/SavoirBundle/Entity/Savoir.php
@@ -386,7 +386,8 @@ class Savoir
      */
     public function removeUpload()
     {
-        if ($this->filenameForRemove) {
+        echo $this->filenameForRemove;
+		if ($this->filenameForRemove && file_exists($this->filenameForRemove)) {
             unlink($this->filenameForRemove);
         }
     }
